+++ src/dot_gen.cpp
@@ -119,7 +119,7 @@ std::string dot_from_dfa(const dfa::Dfa& dfa) {
   out << "digraph g{\n";
   out << "rankdir = \"LR\"\n";
 
-  for (int i = 0; i < dfa.nodes.size(); ++i) {
+  for (size_t i = 0; i < dfa.nodes.size(); ++i) {
     auto& [node, terminal] = dfa.nodes[i];
     
     std::unordered_map<int, std::set<int>> outmap;
